fix(b2b-generator): stawka kandydata z groszami — koniec z zaokrąglaniem 83,5 → 84
W zakładce Kontrakty stawka kandydata wpisana do „szablonu" (Generator umów
B2B) była zaokrąglana do pełnych złotych — recruiter wpisywał 83,5 zł/h
(dokładna kwota trzymana w notatkach wewnętrznych), a na Contract.rate_candidate
lądowało 84,00 zł/h.

Przyczyna: dwie schematy generatora B2B miały `rate_candidate: Optional[int]`,
podczas gdy kolumna `Contract.rate_candidate` to Numeric(12,3), a wszystkie
pozostałe ścieżki stawki (ContractCreate/Update, B2BRenderRequest) są `float`:

- `B2BGenerateRequest` (zapis, POST /generate → contract.rate_candidate =
  payload.rate_candidate): Pydantic NIE zaokrągla — na `int` odrzuca 83.5
  przez 422, więc do szablonu wchodziła ręcznie zaokrąglona liczba całkowita.
- `B2BContractDetailResponse` (odczyt, GET /contracts/{id}/detail): read-back
  stawki ułamkowej ucinał grosze / groził 500 na response_model.

Domknięcie #3e675dc, które naprawiło ten sam błąd 422 tylko na ścieżce
`B2BRenderRequest` (render/DOCX), ale przeoczyło zapis i read-back.

Zmiany:
- oba pola `rate_candidate` int → float (spójne z Numeric(12,3) + resztą stawek),
- testy regresji: fractional 83,5 przechodzi przez oba schematy bez utraty groszy.

Frontend bez zmian — już wysyła `Number("83,5")=83.5` i typuje pole jako `number`.

+++ backend/app/schemas/b2b_contract_generator.py
@@ -68,7 +68,12 @@ class B2BGenerateRequest(BaseModel):
     project_city: Optional[str] = None
     project_description: Optional[str] = None
     correspondence_address: Optional[str] = None
-    rate_candidate: Optional[int] = None
+    # Stawka godzinowa bywa ułamkowa (np. 83,5 PLN/h) — `float`, NIE `int`.
+    # `int` odrzucał 83.5 przez 422 (Pydantic wymusza liczbę całkowitą, NIE
+    # zaokrągla), więc stawka trafiająca na `Contract.rate_candidate`
+    # (Numeric(12,3)) musiała być ręcznie zaokrąglona do 84. Spójne z
+    # `B2BRenderRequest` i `ContractCreate/Update`.
+    rate_candidate: Optional[float] = None
     currency: str = "PLN"
     rate_in_words: Optional[str] = None
     # Nadpisanie zakresu roli na poziomie tej umowy (None = użyj domyślnego).
@@ -94,7 +99,10 @@ class B2BContractDetailResponse(BaseModel):
     project_city: Optional[str] = None
     project_description: Optional[str] = None
     correspondence_address: Optional[str] = None
-    rate_candidate: Optional[int] = None
+    # Odczyt z `Contract.rate_candidate` (Numeric(12,3)) — `float`, by nie ucinać
+    # groszy przy read-backu (i nie wywalać 500 na response_model dla stawki
+    # ułamkowej, np. 83,5).
+    rate_candidate: Optional[float] = None
     currency: Optional[str] = None
     rate_in_words: Optional[str] = None
     scope_items_override: Optional[list[str]] = None
